+++ crates/cheatnet/src/execution/syscalls/cairo1_cheatable_syscall_handler.rs
@@ -1,8 +1,8 @@
-use crate::execution::calls::{execute_inner_call, execute_library_call};
+use crate::execution::calls::cairo1_calls::{execute_inner_call, execute_library_call};
 use crate::execution::contract_print::{contract_print, PrintingResult};
 use crate::execution::entry_point::execute_constructor_entry_point;
 use crate::execution::execution_info::get_cheated_exec_info_ptr;
-use crate::execution::syscalls::lib::stark_felt_from_ptr_immutable;
+use crate::execution::syscalls::stark_felt_from_ptr_immutable;
 use crate::state::CheatnetState;
 use blockifier::abi::constants;
 use blockifier::execution::call_info::CallInfo;
