--- crates/cheatnet/src/execution/syscalls/lib.rs
@@ -1,21 +1,0 @@
-use blockifier::execution::execution_utils::felt_to_stark_felt;
-use cairo_felt::Felt252;
-use cairo_vm::types::relocatable::Relocatable;
-use cairo_vm::vm::errors::vm_errors::VirtualMachineError;
-use cairo_vm::vm::vm_core::VirtualMachine;
-use starknet_api::hash::StarkFelt;
-
-pub fn stark_felt_from_ptr_immutable(
-    vm: &VirtualMachine,
-    ptr: &Relocatable,
-) -> Result<StarkFelt, VirtualMachineError> {
-    Ok(felt_to_stark_felt(&felt_from_ptr_immutable(vm, ptr)?))
-}
-
-pub fn felt_from_ptr_immutable(
-    vm: &VirtualMachine,
-    ptr: &Relocatable,
-) -> Result<Felt252, VirtualMachineError> {
-    let felt = vm.get_integer(*ptr)?.into_owned();
-    Ok(felt)
-}
